Correct morphological processes and photosynthesis
The morphological ratios are constant, only the spacing ratio is returned back to dependency on light availability and flow condition. Solved the issue with negative coral height and plate thickness with time.  The photosynthetic rate is not clipped, therefore in deep zone, where the light does not get, corals are dying. Before they were living forever, that did not make sense.