updated the coral-specific constants read
Set the reading of the coral-specific constants reading procedure, but not sure yet whether it will be functional. Another option is to set them inside the Class Coral, but before I already had the issue that the variables for Coral initialisation have to be pre-defined elsewhere, otherwise it cannot self-initialise. That is why for now I added 2 properties of massive and branching coral constants to the Loop and then in interface I call to read the coral-input files and assign them to later call of Massive and Branching initialisation. it could be wrong though.