Canopy file and core updates
Added function for the light competition into core. Self.light will be only initialised in Coral, the calculations are taking place individually for both goral groupsd in their respective classes. Cleaned up the photosynthesis calculations (The hierarchy is the main question), and the thermal dependency is cleaner now. But, still check whether that "first year"loop was relevant.  After the meeting with Peter also resolved questions in flow. But, the wac calculations and drag are still  a question.